📦 Struct: Rename namespaces
