+++ Common.BasicHelper/Utils/Extensions/Dumpper.cs
@@ -2,7 +2,7 @@
 using System.Net.NetworkInformation;
 using System.Text;
 
-namespace Common.BasicHelper.Util.Extension;
+namespace Common.BasicHelper.Utils.Extensions;
 
 public static class Dumpper
 {
