+++ Common.BasicHelper/Utils/Extensions/ListHelper.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Text;
 
-namespace Common.BasicHelper.Util.Extension;
+namespace Common.BasicHelper.Utils.Extensions;
 
 public static class ListHelper
 {
